Verify that certificate is valid for server hostname
Without this change, the WebSocket library will accept
a trusted certificate issued for domain A when connecting
to domain B. This could be exploited for Man-in-the-middle
attacks.

The underlying issue is that Java considers hostname
verification to be a part of HTTPS and as such, will
not perform it by default.

This change adds the default HostnameVerifier used
by Android, which in recent versions is derived from
OkHttp. Minor changes were made to make it build for
Java 1.6.

Tested with and without a proxy configured.